Time: 23 ms (93.46%) | Memory: 48.6 MB (79.64%) - LeetSync

+++ 560-subarray-sum-equals-k/subarray-sum-equals-k.java
@@ -1,22 +1,21 @@
 import java.util.*;
-
 class Solution {
     public int subarraySum(int[] nums, int k) {
-        HashMap<Long, Long> map = new HashMap<>();
-        long preSum = 0;
-        long cnt = 0;
+        HashMap<Integer,Integer> map = new HashMap<>();
+        int preSum = 0;
+        int cnt = 0;
         //humlog to empty prefix daalna hoga  nhi toh ek count miss hojayega 
-        map.put(0L, 1L);
+        map.put(0, 1);
         for (int x : nums) {
             preSum += x;
-            long need = preSum - k;
+            int need = preSum - k;
             if (map.containsKey(need)) {
                 cnt += map.get(need);
             }
             // we have to store current prefix sum 
-            map.put(preSum, map.getOrDefault(preSum, 0L) + 1);
+            map.put(preSum, map.getOrDefault(preSum, 0) + 1);
         }
 
-        return (int) cnt;
+        return cnt;
     }
 }
